requirements.txt
Flask==2.3.7
line-bot-sdk==2.3.0